move loading spinner to common folder
