+++ src/components/common/loading/Spinner.vue
@@ -1,7 +1,7 @@
 <template>
     <div class="mt-10 items-center justify-center flex flex-col">
         <h2 class="text-3xl text-gray-800 mb-3">Loading Contents...</h2>
-        <img src="../../assets/loading-spinner.svg" class="" />
+        <img src="../../../assets/loading-spinner.svg" class="" />
     </div>
 </template>
