+++ src/components/profile/Wrapper.vue
@@ -11,11 +11,11 @@
 </template>
 
 <script lang="ts">
-    import PrivateProfile from "../../views/common/PrivateProfile.vue";
-    import PublicProfile from "../../views/common/PublicProfile.vue";
+    import PrivateProfile from "@/views/common/PrivateProfile.vue";
+    import PublicProfile from "@/views/common/PublicProfile.vue";
     import { store } from "@/store/store";
     import { Role } from "@/entities/Role";
-    import LoadingComponent from "../../components/loading/Spinner.vue";
+    import LoadingComponent from "@/components/common/loading/Spinner.vue";
     import { checkPrivilege } from "@/use/PermissionHelper";
 
     export default {
