+++ src/views/admin/EditCreateAccountForm.vue
@@ -137,13 +137,13 @@
     import LecturerEntity from "@/entities/LecturerEntity";
     import AdminEntity from "@/entities/AdminEntity";
     import { Account } from "@/entities/Account";
-    import Admin from "../../api/api_models/user_management/Admin";
-    import Student from "../../api/api_models/user_management/Student";
-    import Lecturer from "../../api/api_models/user_management/Lecturer";
+    import Admin from "@/api/api_models/user_management/Admin";
+    import Student from "@/api/api_models/user_management/Student";
+    import Lecturer from "@/api/api_models/user_management/Lecturer";
     import DeleteAccountModal from "@/components/modals/DeleteAccountModal.vue";
     import { Country } from "@/entities/Country";
     import ErrorBag from "@/use/ErrorBag";
-    import ValidationResponseHandler from "../../use/ValidationResponseHandler";
+    import ValidationResponseHandler from "@/use/ValidationResponseHandler";
     import GenericResponseHandler from "@/use/GenericResponseHandler";
     import BirthDatePicker from "@/components/BirthDatePicker.vue";
     import RoleSection from "@/components/account/edit/sections/RoleSection.vue";
@@ -151,7 +151,7 @@
     import PersonalInformationSection from "@/components/account/edit/sections/PersonalInformationSection.vue";
     import LecturerInformationSection from "@/components/account/edit/sections/LecturerInformationSection.vue";
     import StudentInformationSection from "@/components/account/edit/sections/StudentInformationSection.vue";
-    import LoadingComponent from "../../components/loading/Spinner.vue";
+    import LoadingComponent from "@/components/common/loading/Spinner.vue";
     import { checkPrivilege } from "@/use/PermissionHelper";
     import UnsavedChangesModal from "@/components/modals/UnsavedChangesModal.vue";
     import { onBeforeRouteUpdate, onBeforeRouteLeave } from "vue-router";
