+++ src/views/common/Welcome.vue
@@ -26,7 +26,7 @@
 <script lang="ts">
     import { useStore } from "@/store/store";
     import { ref, onBeforeMount } from "vue";
-    import LoadingSpinner from "@/components/loading/Spinner.vue";
+    import LoadingSpinner from "@/components/common/loading/Spinner.vue";
     import { checkPrivilege } from "@/use/PermissionHelper";
     import { Role } from "@/entities/Role";
